pyproject.toml
[project]
name = "repl-nix-dataingestpromptquery"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=0.5.18",
    "langchain>=0.3.7",
    "openai>=1.54.3",
    "pillow>=11.0.0",
    "psycopg2-binary>=2.9.10",
    "pymupdf>=1.24.13",
    "python-docx>=1.1.2",
    "streamlit>=1.40.0",
]
